% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFFparser.R
\name{GFFparser}
\alias{GFFparser}
\title{Parse GFF File and Extract Attributes}
\usage{
GFFparser(file)
}
\arguments{
\item{file}{The path to the GFF file.}
}
\value{
A data frame with the GFF content and extracted attributes.
}
\description{
This function reads a GFF file, extracts attributes, and returns a dataframe with the attributes.
}
